frontend-service/requirements.txt
flask-socketio
flask-testing
pytest
requests